agentic_rag/requirements.txt
# Azure AI Foundry Agentic RAG System Dependencies

# Azure AI Foundry SDK
azure-ai-projects>=1.0.0b10
azure-ai-inference>=1.0.0b9
azure-ai-agents>=1.0.0b10
azure-identity>=1.15.0
azure-core>=1.30.0

# Azure AI Search for vector storage
azure-search-documents>=11.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
requests>=2.31.0

# Development tools
pytest>=7.4.0

# UI Framework
gradio>=4.0.0
